+++ goeffel/analysis.py
@@ -69,6 +69,11 @@ COLUMN_PLOT_CONFIGS = {
         'plot_title': 'foo',
         'rolling_wdw_width_seconds': 0,
     },
+    'proc_mem_rss_percent': {
+         'y_label': 'Process RSS mem [%]',
+         'plot_title': 'foo',
+         'rolling_wdw_width_seconds': 5
+    },
     'system_loadavg1': {
         'y_label': 'System 1 min load avg',
         'plot_title': 'foo',